refactor(cli): extract the shared job core to break the import cycle [ROBO-5779]
Move the transport-agnostic job core (server state, command table, arg
parsing, and the isolated command runner) out of cli_server into a new
_server_core module. Both the HTTP (cli_server) and uipath-ipc
(cli_server_ipc) transports now import it at module top level, so
cli_server_ipc no longer needs its function-local import of cli_server and
the two modules no longer form an import cycle.

Also switch the Windows branch of _run_server to
asyncio.Runner(loop_factory=ProactorEventLoop): it runs the server on the
Proactor loop without mutating the thread's current-loop slot and tears
tasks down cleanly on exit.
