+++ packages/uipath/src/uipath/_cli/cli_server.py
@@ -2,7 +2,6 @@ import asyncio
 import importlib
 import json
 import os
-import shlex
 import sys
 import tempfile
 import time
@@ -13,11 +12,14 @@ from typing import Any
 import click
 from aiohttp import ClientSession, UnixConnector, web
 
+from ._server_core import (
+    COMMANDS,
+    _run_command_isolated,
+    _state,
+    parse_args,
+)
 from ._telemetry import track_command
 from ._utils._console import ConsoleLogger
-from .cli_debug import debug
-from .cli_eval import eval
-from .cli_run import run
 from .cli_server_ipc import (
     IPythonRuntimeServer,
     PythonRunRequest,
@@ -42,30 +44,6 @@ IS_WINDOWS = sys.platform == "win32"
 SOCKET_ENV_VAR = "UIPATH_SERVER_SOCKET"
 DEFAULT_SOCKET_PATH = "/tmp/uipath-server.sock"
 DEFAULT_PORT = 8765
-
-COMMANDS = {
-    "run": run,
-    "debug": debug,
-    "eval": eval,
-}
-
-
-class _ServerState:
-    """Mutable server state, initialized lazily at server startup."""
-
-    def __init__(self) -> None:
-        self.lock: asyncio.Lock | None = None
-        self.baseline_env: dict[str, str] | None = None
-
-    def init(self) -> None:
-        """Must be called inside a running event loop at server startup."""
-        if self.lock is not None:
-            return
-        self.lock = asyncio.Lock()
-        self.baseline_env = os.environ.copy()
-
-
-_state = _ServerState()
 
 
 DEFAULT_PRELOAD_MODULES = [
@@ -123,80 +101,6 @@ def get_field(message: dict[str, Any], *keys: str) -> Any:
         if key in message:
             return message[key]
     return None
-
-
-def parse_args(args: str | list[str] | None) -> list[str]:
-    """Parse args into a list of strings."""
-    if args is None:
-        return []
-    if isinstance(args, list):
-        return args
-    if isinstance(args, str):
-        return shlex.split(args)
-    return []
-
-
-async def _run_command_isolated(
-    cmd: Any,
-    args: list[str],
-    env_vars: dict[str, str],
-    working_dir: str | None,
-) -> dict[str, Any]:
-    """Run one command with per-job env/cwd isolation (the shared job core)."""
-    if _state.lock is None or _state.baseline_env is None:
-        raise RuntimeError("Server state not initialized")
-
-    async with _state.lock:
-        original_cwd = os.getcwd()
-        try:
-            # Start from server baseline + request env vars only, so nothing from
-            # a previous job leaks through.
-            os.environ.clear()
-            os.environ.update(_state.baseline_env)
-            if isinstance(env_vars, dict):
-                os.environ.update(env_vars)
-
-            if working_dir and isinstance(working_dir, str):
-                try:
-                    os.chdir(working_dir)
-                except (FileNotFoundError, NotADirectoryError, PermissionError) as e:
-                    # Request-shaped error: the caller gave a bad working dir.
-                    # HTTP surfaces this as 400; IPC just returns ExitCode/Error.
-                    return {
-                        "ExitCode": 1,
-                        "Error": f"Cannot change to working directory: {e}",
-                        "Result": None,
-                        "Unexpected": False,
-                        "ClientError": True,
-                    }
-
-            result_value = await asyncio.to_thread(
-                cmd.main, args, standalone_mode=False
-            )
-            return {
-                "ExitCode": 0,
-                "Error": None,
-                "Result": result_value,
-                "Unexpected": False,
-            }
-        except SystemExit as e:
-            exit_code = e.code if isinstance(e.code, int) else 1
-            return {
-                "ExitCode": exit_code,
-                "Error": None if exit_code == 0 else f"Exit code: {exit_code}",
-                "Result": None,
-                "Unexpected": False,
-            }
-        except Exception as e:  # report any job failure as a result, not a fault
-            return {"ExitCode": 1, "Error": str(e), "Result": None, "Unexpected": True}
-        finally:
-            # Restore to server baseline.
-            try:
-                os.chdir(original_cwd)
-            except OSError:
-                pass
-            os.environ.clear()
-            os.environ.update(_state.baseline_env)
 
 
 # --------------------------------------------------------------------------- #
@@ -479,16 +383,9 @@ def _run_server(
     )
     coro = _serve(ack_socket_path, server_socket, port or DEFAULT_PORT, use_tcp)
     try:
-        # Windows named pipes need the Proactor loop; build it explicitly since another
-        # lib (e.g. socketio) may have flipped the policy to Selector. Gate on the
-        # sys.platform literal so mypy narrows ProactorEventLoop (Windows-only) here.
         if sys.platform == "win32":
-            loop = asyncio.ProactorEventLoop()
-            asyncio.set_event_loop(loop)
-            try:
-                loop.run_until_complete(coro)
-            finally:
-                loop.close()
+            with asyncio.Runner(loop_factory=asyncio.ProactorEventLoop) as runner:
+                runner.run(coro)
         else:
             asyncio.run(coro)
     except KeyboardInterrupt:
